refactor: split monolithic callback registration into focused modules
Alert callbacks (479 LOC -> 3 sub-modules):
- alert_preferences.py: 7 preference management callbacks
- alert_history.py: 4 alert history callbacks
- alert_notifications.py: status indicator + 5 clientside callbacks
- alert_models.py: AlertPreferences Pydantic model replacing paired
  build_preferences_dict/extract_preferences_tuple functions

Dashboard callbacks (460 LOC -> 3 sub-modules):
- dashboard_callbacks/period.py: period button selection + countdown
- dashboard_callbacks/content.py: main update + post feed callbacks
- dashboard_callbacks/table.py: screener row click, filter, expand

Both alerts.py and dashboard.py reduced to thin orchestrators that
delegate to the focused sub-modules.
