+++ shitty_ui/callbacks/alert_preferences.py
@@ -1,0 +1,155 @@
+"""Preference management callbacks for the alert system.
+
+Handles:
+- Opening/closing the alert config panel
+- Showing/hiding conditional inputs (email, SMS, quiet hours)
+- Confidence threshold display
+- Saving preferences to localStorage
+- Loading preferences from localStorage into form
+"""
+
+from dash import Dash, Input, Output, State, no_update
+
+from callbacks.alert_models import AlertPreferences
+from data import get_available_assets
+
+
+def register_alert_preference_callbacks(app: Dash) -> None:
+    """Register all preference-management alert callbacks.
+
+    Args:
+        app: The Dash application instance.
+    """
+
+    @app.callback(
+        Output("alert-config-offcanvas", "is_open"),
+        [Input("open-alert-config-button", "n_clicks")],
+        [State("alert-config-offcanvas", "is_open")],
+        prevent_initial_call=True,
+    )
+    def toggle_alert_config(n_clicks: int, is_open: bool) -> bool:
+        """Toggle the alert configuration panel open/closed."""
+        if n_clicks:
+            return not is_open
+        return is_open
+
+    @app.callback(
+        Output("email-input-collapse", "is_open"),
+        [Input("alert-email-toggle", "value")],
+    )
+    def toggle_email_input(email_enabled: bool) -> bool:
+        """Show email input when email alerts are enabled."""
+        return bool(email_enabled)
+
+    @app.callback(
+        Output("sms-input-collapse", "is_open"),
+        [Input("alert-sms-toggle", "value")],
+    )
+    def toggle_sms_input(sms_enabled: bool) -> bool:
+        """Show SMS input when SMS alerts are enabled."""
+        return bool(sms_enabled)
+
+    @app.callback(
+        Output("quiet-hours-collapse", "is_open"),
+        [Input("alert-quiet-hours-toggle", "value")],
+    )
+    def toggle_quiet_hours(quiet_enabled: bool) -> bool:
+        """Show quiet hours time inputs when quiet hours are enabled."""
+        return bool(quiet_enabled)
+
+    @app.callback(
+        Output("confidence-threshold-display", "children"),
+        [Input("alert-confidence-slider", "value")],
+    )
+    def update_confidence_display(value: float) -> str:
+        """Display the confidence slider value as a percentage."""
+        return f"{int(value * 100)}%"
+
+    @app.callback(
+        Output("alert-assets-dropdown", "options"),
+        [Input("alert-config-offcanvas", "is_open")],
+    )
+    def populate_alert_assets(is_open: bool):
+        """Populate the assets dropdown when the panel opens."""
+        if not is_open:
+            return no_update
+        available_assets = get_available_assets()
+        return [{"label": asset, "value": asset} for asset in available_assets]
+
+    @app.callback(
+        [
+            Output("alert-preferences-store", "data"),
+            Output("alert-save-toast", "is_open"),
+        ],
+        [Input("save-alert-prefs-button", "n_clicks")],
+        [
+            State("alert-master-toggle", "value"),
+            State("alert-confidence-slider", "value"),
+            State("alert-assets-dropdown", "value"),
+            State("alert-sentiment-radio", "value"),
+            State("alert-browser-toggle", "value"),
+            State("alert-email-toggle", "value"),
+            State("alert-email-input", "value"),
+            State("alert-sms-toggle", "value"),
+            State("alert-sms-input", "value"),
+            State("alert-quiet-hours-toggle", "value"),
+            State("quiet-hours-start", "value"),
+            State("quiet-hours-end", "value"),
+        ],
+        prevent_initial_call=True,
+    )
+    def save_alert_preferences(
+        n_clicks: int,
+        enabled: bool,
+        min_confidence: float,
+        assets: list,
+        sentiment: str,
+        browser_on: bool,
+        email_on: bool,
+        email_addr: str,
+        sms_on: bool,
+        sms_phone: str,
+        quiet_on: bool,
+        quiet_start: str,
+        quiet_end: str,
+    ):
+        """Gather all form values and write them to the localStorage-backed store."""
+        preferences = AlertPreferences.from_form_fields(
+            enabled,
+            min_confidence,
+            assets,
+            sentiment,
+            browser_on,
+            email_on,
+            email_addr,
+            sms_on,
+            sms_phone,
+            quiet_on,
+            quiet_start,
+            quiet_end,
+        )
+        return preferences.to_dict(), True
+
+    @app.callback(
+        [
+            Output("alert-master-toggle", "value"),
+            Output("alert-confidence-slider", "value"),
+            Output("alert-assets-dropdown", "value"),
+            Output("alert-sentiment-radio", "value"),
+            Output("alert-browser-toggle", "value"),
+            Output("alert-email-toggle", "value"),
+            Output("alert-email-input", "value"),
+            Output("alert-sms-toggle", "value"),
+            Output("alert-sms-input", "value"),
+            Output("alert-quiet-hours-toggle", "value"),
+            Output("quiet-hours-start", "value"),
+            Output("quiet-hours-end", "value"),
+        ],
+        [Input("alert-config-offcanvas", "is_open")],
+        [State("alert-preferences-store", "data")],
+    )
+    def load_preferences_into_form(is_open: bool, prefs: dict):
+        """When the offcanvas opens, populate form fields from stored preferences."""
+        if not is_open or not prefs:
+            return (no_update,) * 12
+        return AlertPreferences.from_stored(prefs).to_form_tuple()
