+++ shitty_ui/pages/dashboard_callbacks/__init__.py
@@ -1,0 +1,17 @@
+"""Dashboard callback sub-modules.
+
+Each module registers a focused group of callbacks:
+- period: Time period selection and countdown
+- content: Main dashboard content (KPIs, screener, insights, post feed)
+- table: Data table management (collapse, filtering, row clicks, thesis expand)
+"""
+
+from pages.dashboard_callbacks.period import register_period_callbacks
+from pages.dashboard_callbacks.content import register_content_callbacks
+from pages.dashboard_callbacks.table import register_table_callbacks
+
+__all__ = [
+    "register_period_callbacks",
+    "register_content_callbacks",
+    "register_table_callbacks",
+]
